Make CLI reusable between binaries
